# src/main/resources/swagger.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Flight & Airport API",
    "description": "Flight and Airport CRUD and search API.",
    "contact": {
      "name": "Cemre Şenyuva",
      "url": "https://cemrecodes.github.io",
      "email": "cemressenyuva@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
    "/airports": {
      "get": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Get all airports/get airport by city",
        "description": "get-airport",
        "operationId": "getCollectionResource-airport-get_1_1",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all airports/ Airport by given city exists and is returned.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelAirport"
                }
              },
              "application/x-spring-data-compact+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelAirport"
                }
              },
              "text/uri-list": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Airport doesn't exist",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Save airport",
        "description": "create-airport",
        "operationId": "postCollectionResource-airport-post_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Airport is successfully created.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelAirport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Airport already exists.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/airports/{airportId}": {
      "get": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Get airport by ID",
        "operationId": "getAirport",
        "parameters": [
          {
            "name": "airportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Airport exists and is returned.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "404": {
            "description": "Airport doesn't exist",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Update airport by ID",
        "operationId": "updateAirport",
        "parameters": [
          {
            "name": "airportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Airport has been updated successfully.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Delete airport by ID",
        "operationId": "deleteAirport",
        "parameters": [
          {
            "name": "airportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Airport has been successfully deleted."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/airports/addAirportWithFaker": {
      "get": {
        "tags": [
          "Airport Controller"
        ],
        "summary": "Add mock airport",
        "operationId": "addAirportWithFaker",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Airport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string"
          }
        }
      },
      "EntityModelAirport": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "PagedModelEntityModelAirport": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "airports": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelAirport"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "AirportDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string"
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}